requirements.txt
certifi==2021.10.8
cfdiclient==1.4.2
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
cryptography==36.0.1
Flask==2.0.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
lxml==4.7.1
MarkupSafe==2.0.1
pycparser==2.21
pycryptodome==3.14.1
pyOpenSSL==22.0.0
python-dateutil==2.8.2
requests==2.27.1
six==1.16.0
urllib3==1.26.8
Werkzeug==2.0.2
